+++ src/main/java/infrastructure/date/SystemDateProvider.java
@@ -1,4 +1,4 @@
-package infrastructure;
+package infrastructure.date;
 
 import domain.DateProvider;
 import java.time.LocalDate;
